Update to de_DE.md
I started with the translation of the actual Text. Had to make small grammatical changes to make it read more natural but all in all tried to stay as close to the original as possible. Still, a second opinion would be great of course. I also changed the asterisks for gendering to underscores in order to hopefully prevent unwanted Markdown formatting. :)
Feedback and changes are always appreciated!